made LineEditBox. Also changed structure of Application_Level

+++ view/Font.cpp
@@ -9,7 +9,7 @@
 
 namespace view {
 
-    Font::Font(const std::string& path, Uint16 fontSize)
+    Font::Font(const std::string& path, Uint32 fontSize)
         : m_font(TTF_OpenFont(path.c_str(), fontSize)), m_loadedCorrectly(m_font != nullptr) {
         if (m_font == nullptr) {
             std::cout << "Could not load font! SDL_ttf Error: " << TTF_GetError() << '\n';
